refactor club ot DDD

+++ EL-t3.Domain/Entities/PlayerSeason.cs
@@ -2,7 +2,7 @@ namespace EL_t3.Domain.Entities;
 public class PlayerSeason
 {
     public int Id { get; set; }
-    public int ClubId { get; set; }
+    public long ClubId { get; set; }
     public Club? Club { get; set; }
     public int PlayerId { get; set; }
     public Player? Player { get; set; }
